pyproject.toml
[tool.poetry]
name = "chocopkgupdater"
version = "0.1.0"
description = ""
authors = ["Tom Emv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
requests = "^2.31.0"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.3"
jinja2 = "^3.1.3"
lxml = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
